// cpo-server/src/types/Tag.ts
import CreatedUpdatedProps from './CreatedUpdatedProps';
import { ImportStatus } from './GlobalType';
import { OCPIToken } from './ocpi/OCPIToken';
import { TagAuthorizationActions } from './Authorization';
import User from './User';

export default interface Tag extends CreatedUpdatedProps, TagAuthorizationActions {
  id: string;
  description?: string;
  visualID?: string;
  issuer: boolean;
  active: boolean;
  userID?: string;
  transactionsCount?: number;
  ocpiToken?: OCPIToken;
  user?: User;
  default?: boolean,
  fromEMSP?:boolean,
  importedData?: {
    autoActivateUserAtImport: boolean;
    autoActivateTagAtImport: boolean;
  };
}

export interface ImportedTag {
  id: string;
  visualID: string;
  description: string;
  importedBy?: string;
  importedOn?: Date;
  status?: ImportStatus;
  errorDescription?: string;
  name?: string;
  firstName?: string;
  email?: string;
  fromEMSP?:boolean,
  importedData?: {
    autoActivateUserAtImport: boolean;
    autoActivateTagAtImport: boolean;
  };
  siteIDs?: string;
}

export const TagRequiredImportProperties = [
  'id'
];